requirements.txt
websocket-client
pandas
python-dotenv